// src/app/lib/types.ts
export interface glAccountType {
  idGlAccount?: number;
  uniqueId?: string;
  glAccount: string;
  description?: string;
  groupDetail: string;
  groupGl: string;
  sap: boolean;
  active: boolean;
  createdAt?: Date;
  createdBy?: string;
  updatedAt?: Date;
  updatedBy?: string | null;
}

export interface mGlAccountType {
  idGlAccount: number;
  glAccount: string;
  groupGL: string;
  groupDetail: string;
}
export interface mCostCenterType {
  idCostCenter: number;
  costCenter: string;
  bidang: string;
  dinas: string;
}
export interface availableType {
  available: number;
  mGlAccount: mGlAccountType;
  mCostCenter: mCostCenterType;
}
export interface selectType {
  id: string | number;
  value: string;
}
// ...
export interface RKAPType {
  [x: string]: any;
  title: string;
  total: number;
  month: {
    JANUARI: number;
    FEBRUARI: number;
    MARET: number;
    APRIL: number;
    MEI: number;
    JUNI: number;
    JULI: number;
    AGUSTUS: number;
    SEPTEMBER: number;
    OKTOBER: number;
    NOVEMBER: number;
    DESEMBER: number;
  };
  groupDetail?: RKAPGroupDetailType[];
}
export interface ViewBudgetMasterData {
  budget: RKAPType | null;
  actual: RKAPType | null;
  remaining: RKAPType | null;
}
export interface BudgetMTDYTD {
  budget: { BudgetMTD: number; BudgetYTD: number };
}
export interface RKAPGroupDetailType {
  title: string;
  total: number;
  glNumber: number;
  month: {
    JANUARI: number;
    FEBRUARI: number;
    MARET: number;
    APRIL: number;
    MEI: number;
    JUNI: number;
    JULI: number;
    AGUSTUS: number;
    SEPTEMBER: number;
    OKTOBER: number;
    NOVEMBER: number;
    DESEMBER: number;
  };
}

export interface CostCenterType {
  idCostCenter?: number;
  uniqueId?: string;
  costCenter: string;
  description: string;
  bidang: string;
  dinas: string;
  directorat: string;
  groupDinas: string;
  profitCenter: string;
  active: boolean;
  createdAt?: Date;
  createdBy?: string;
  updatedAt?: Date;
  updatedBy?: string | null;
}